+++ src/main/java/com/hoodee/community/Controlller/QuestionController.java
@@ -22,7 +22,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Long id,
+    public String question(@PathVariable(name = "id") Integer id,
                            Model model){
 
         QuestionDTO questionDTO = questionService.getByID(id);
